ul-30 Updated api structure
Moved endpoint view functions into their own modules namespaced
by their system function.
